feat: Bump Mongo driver and Spring Boot versions

+++ src/main/kotlin/io/lsdconsulting/lsd/distributed/mongo/config/LibraryConfig.kt
@@ -34,7 +34,7 @@ open class LibraryConfig {
         @Value("\${lsd.dist.connectionString}") dbConnectionString: String,
         @Value("\${lsd.dist.db.trustStoreLocation:#{null}}") trustStoreLocation: String?,
         @Value("\${lsd.dist.db.trustStorePassword:#{null}}") trustStorePassword: String?,
-        @Value("\${lsd.dist.db.connectionTimeout.millis:#{" + DEFAULT_TIMEOUT_MILLIS + "}}") connectionTimeout: Int,
+        @Value("\${lsd.dist.db.connectionTimeout.millis:#{" + DEFAULT_TIMEOUT_MILLIS + "}}") connectionTimeout: Long,
         @Value("\${lsd.dist.db.collectionSizeLimit.megabytes:#{" + DEFAULT_COLLECTION_SIZE_LIMIT_MBS + "}}") collectionSizeLimit: Long,
     ): InterceptedInteractionCollectionBuilder = InterceptedInteractionCollectionBuilder(
         dbConnectionString,
